Made PlayMMFragment show up
Removed PlayActivity, logic was moved to GameActivity and
PlayMenuFragment.
Will still need to make login/signup/[any current activity thats not
play related] as a fragment in MenuActivity, and make anything play
related a fragment in GameActivity.
Buttons in PlayMM are transparent, but may not matter if we change
game/ui.